+++ paper/src/main/java/com/badbones69/crazyenchantments/paper/enchantments/MaceEnchantments.java
@@ -1,0 +1,62 @@
+package com.badbones69.crazyenchantments.paper.enchantments;
+
+import com.badbones69.crazyenchantments.paper.CrazyEnchantments;
+import com.badbones69.crazyenchantments.paper.Methods;
+import com.badbones69.crazyenchantments.paper.Starter;
+import com.badbones69.crazyenchantments.paper.api.enums.CEnchantments;
+import com.badbones69.crazyenchantments.paper.api.objects.CEnchantment;
+import com.badbones69.crazyenchantments.paper.api.utils.EnchantUtils;
+import com.badbones69.crazyenchantments.paper.api.utils.EventUtils;
+import com.badbones69.crazyenchantments.paper.controllers.settings.EnchantmentBookSettings;
+import com.badbones69.crazyenchantments.paper.support.PluginSupport;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageByEntityEvent;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.bukkit.potion.PotionEffect;
+import org.bukkit.potion.PotionEffectType;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Map;
+
+public class MaceEnchantments implements Listener {
+
+    @NotNull
+    private final CrazyEnchantments plugin = JavaPlugin.getPlugin(CrazyEnchantments.class);
+
+    @NotNull
+    private final Starter starter = this.plugin.getStarter();
+
+    @NotNull
+    private final Methods methods = this.starter.getMethods();
+
+    @NotNull
+    private final EnchantmentBookSettings enchantmentBookSettings = this.starter.getEnchantmentBookSettings();
+
+    // Plugin Support.
+    @NotNull
+    private final PluginSupport pluginSupport = this.starter.getPluginSupport();
+
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    public void onPlayerDamage(EntityDamageByEntityEvent event) {
+        if (EventUtils.isIgnoredEvent(event)) return;
+        if (this.pluginSupport.isFriendly(event.getDamager(), event.getEntity())) return;
+
+        if (!(event.getEntity() instanceof LivingEntity entity)) return;
+        if (!(event.getDamager() instanceof Player damager)) return;
+
+        ItemStack item = this.methods.getItemInHand(damager);
+
+        if (entity.isDead()) return;
+
+        Map<CEnchantment, Integer> enchantments = this.enchantmentBookSettings.getEnchantments(item);
+
+        if (EnchantUtils.isEventActive(CEnchantments.HIGHRISE, damager, item, enchantments)) {
+            damager.addPotionEffect(new PotionEffect(PotionEffectType.JUMP_BOOST, (enchantments.get(CEnchantments.HIGHRISE.getEnchantment()) + 5) * 20, 3));
+        }
+    }
+}