+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-### Changes:
-- Simple changes and Build to test
+# Changes:
+### FEW VERSIONS NOW -
+Added the following:
+- Nest enchant for armor
+- Roost enchant for armor
+- Selki enchant for boots
+
+Added a mace enchant class!
+- Highrise enchant for maces
+- Enchantment info